plugins/audiofilters/requirements.txt
numpy
wave
openpyxl
matplotlib
pyyaml
pydub
